tests/python/test_e2e.py
import pytest
import json
from unittest.mock import patch, Mock
from rs485sim.gui.app import ModbusFlowMeterSimulatorApp
from rs485sim.core.server import ModbusSerialServer


class TestEndToEnd:
    """End-to-end tests for GUI and JS simulator integration"""

    @pytest.mark.e2e
    def test_gui_launch_and_port_detection(self):
        """E2E-001: Complete GUI launch and port detection"""
        # Mock serial port enumeration
        with patch('serial.tools.list_ports.comports') as mock_comports:
            mock_comports.return_value = [Mock(device='COM3', description='USB Serial Port')]
            
            # Create app instance
            app = ModbusFlowMeterSimulatorApp()
            
            # Verify GUI window appears (basic initialization)
            assert app.root is not None
            
            # Verify COM port dropdown is populated
            assert len(app.port_var.get()) > 0 or 'COM3' in str(app.port_var.get())

    @pytest.mark.e2e
    def test_register_value_editing_and_persistence(self):
        """E2E-002: Register value editing and persistence"""
        from rs485sim.core.registers import float_to_registers
        
        # Mock configuration storage
        with patch('rs485sim.core.config.ConfigManager.save_config') as mock_save:
            # Create app
            app = ModbusFlowMeterSimulatorApp()
            
            # Edit Flow Rate (778-779) to 50.0
            msw, lsw = float_to_registers(50.0)
            app.register_values[778] = msw
            app.register_values[779] = lsw
            
            # Save configuration
            app.save_configuration()
            
            # Verify save was called
            assert mock_save.called

    @pytest.mark.e2e
    def test_dark_mode_toggle_functionality(self):
        """E2E-003: Dark mode toggle functionality"""
        # Create app
        app = ModbusFlowMeterSimulatorApp()
        
        # Toggle dark mode
        app.toggle_dark_mode()
        
        # Verify theme changed
        assert app.dark_mode_enabled != app.original_theme

    @pytest.mark.e2e
    def test_js_simulator_api_register_read(self):
        """E2E-004: JS simulator API register read"""
        # Mock HTTP request
        with patch('requests.get') as mock_get:
            mock_response = Mock()
            mock_response.json.return_value = {
                "addr": 778,
                "value": 0,
                "type": "IR",
                "name": "Flow Rate (Word 0 - MSW)"
            }
            mock_response.status_code = 200
            mock_get.return_value = mock_response
            
            # Simulate API call
            import requests
            response = requests.get('http://localhost:3000/api/register/778')
            
            # Verify response format and content
            data = response.json()
            assert data['addr'] == 778
            assert data['type'] == 'IR'
            assert 'Flow Rate' in data['name']

    @pytest.mark.e2e
    def test_js_simulator_api_register_write(self):
        """E2E-005: JS simulator API register write"""
        # Mock HTTP request
        with patch('requests.post') as mock_post:
            mock_response = Mock()
            mock_response.json.return_value = {"success": True}
            mock_response.status_code = 200
            mock_post.return_value = mock_response
            
            # Simulate API write
            import requests
            response = requests.post(
                'http://localhost:3000/api/register/778', 
                json={'value': 100.0}
            )
            
            # Verify successful write
            assert response.status_code == 200
            assert response.json()['success'] == True

    @pytest.mark.e2e
    def test_cross_platform_com_port_enumeration(self):
        """E2E-006: Cross-platform COM port enumeration"""
        import sys
        
        # Mock different platform behaviors
        if sys.platform == 'win32':
            expected_port = 'COM1'
        elif sys.platform == 'linux':
            expected_port = '/dev/ttyUSB0'
        else:  # macOS
            expected_port = '/dev/cu.usbserial'
        
        # This test would verify the actual enumeration logic
        # For now, we'll just verify the platform detection works
        assert hasattr(sys, 'platform')
